declare TokenTag in types file
