--- src/tags.ts
@@ -1,9 +1,0 @@
-export enum TokenTag {
-  VAULT = "vault",
-  ASSET_PROXY = "assetproxy",
-  CCPOOL = "ccpool",
-  PRINCIPAL = "eP",
-  WPOOL = "wpool",
-  YIELD = "eY",
-  CURVE = "curve",
-}
